chore(aep-api): remove deprecated requirements/design HTTP surface
get-requirements, discard-requirements, list-requirements-versions,
get-requirements-at-version, get-design-bundle, get-design-bundle-at-tag,
discard-design-changes, list-design-versions, and collect-dependency-spec
were already tagged (Deprecated) and unused by apps/console — spec/design
content is read via the Files API (list-files/read-file) instead. Removed
the 9 operations and cascaded through everything that only existed to back
them: requirements_service.go and its whole feature package (collab-session
stays, on its own file); the design-side GetDesign/GetDesignBundle/
DiscardChanges/ListDesignVersions/decodeDesignTag chain (SaveAndProceed,
CollectSpec, and MarkOrgPublished stay — provisioning and the save-gate
still call them); ArtifactService.DiscardRequirements/DiscardDesign and
their revert-commit helper; the now-orphaned RequirementsBundle model; and
a stale internal/arch allowlist edge.
