+++ services/aep-api/internal/api/huma_register.go
@@ -27,7 +27,6 @@ import (
 	"github.com/wso2/aep/aep-api/internal/feature/artifacts"
 	"github.com/wso2/aep/aep-api/internal/feature/build"
 	"github.com/wso2/aep/aep-api/internal/feature/component"
-	"github.com/wso2/aep/aep-api/internal/feature/design"
 	"github.com/wso2/aep/aep-api/internal/feature/execution"
 	"github.com/wso2/aep/aep-api/internal/feature/files"
 	"github.com/wso2/aep/aep-api/internal/feature/genai"
@@ -53,9 +52,7 @@ type HumaDeps struct {
 	OrgSvc            organization.OrganizationService
 	ComponentSvc      component.ComponentService
 	ConfigSvc         component.ConfigService
-	RequirementsSvc   requirements.RequirementsService
 	CollabRepo        gitrepo.RepoService
-	DesignSvc         design.DesignService
 	ProvisioningSvc   *provisioning.Service
 	TaskReads         *task.Reads
 	TaskCommands      *task.Commands
@@ -89,9 +86,7 @@ func RegisterAllHuma(api huma.API, d HumaDeps) {
 	organization.RegisterOrganization(api, d.OrgSvc)
 	component.RegisterComponent(api, d.ComponentSvc)
 	component.RegisterConfig(api, d.ConfigSvc)
-	requirements.RegisterRequirements(api, d.RequirementsSvc)
 	requirements.RegisterCollab(api, d.CollabRepo)
-	design.RegisterDesign(api, d.DesignSvc)
 	provisioning.RegisterResources(api, d.ProvisioningSvc)
 	task.RegisterTask(api, d.TaskReads, d.TaskCommands, d.TaskPlan)
 	execution.RegisterTaskStream(api, d.TaskStream)
