+++ services/aep-api/internal/api/huma_registration_test.go
@@ -35,7 +35,7 @@ func TestHumaRegistration_NoDupAndComplete(t *testing.T) {
 
 	wantOps := []string{
 		"list-projects", "list-organizations", "list-components", "get-component-config",
-		"get-requirements", "get-spec-collab-session", "get-design-bundle",
+		"get-spec-collab-session",
 		// Tasks-github-native surface (§9.1): the board/dispatch/retry ops are gone.
 		"list-tasks", "get-task", "plan-tasks", "execute-task", "hold-task", "unhold-task",
 		// The task-log endpoint is now one SSE stream (status + executions +
@@ -58,6 +58,11 @@ func TestHumaRegistration_NoDupAndComplete(t *testing.T) {
 	wantAbsent := []string{
 		"start-devflow", "list-devflows", "get-devflow", "decide-devflow-gate",
 		"save-requirements", "save-design",
+		// Requirements/design deprecated read+version surface — superseded by
+		// the Files API (list-files/read-file); removed outright.
+		"get-requirements", "discard-requirements", "list-requirements-versions", "get-requirements-at-version",
+		"get-design-bundle", "discard-design-changes", "list-design-versions", "get-design-bundle-at-tag",
+		"collect-dependency-spec",
 	}
 	for _, op := range wantAbsent {
 		if strings.Contains(s, op) {
