+++ services/aep-api/internal/feature/artifacts/save_ops.go
@@ -16,21 +16,19 @@
 
 package artifacts
 
-// The Workspace-port write primitives behind Save (annotated tag at the pinned
-// commit) and Discard (one revert Mutate rewriting a subtree back to its last
-// tag). Save creates no commit — the accepted draft is already on `main` — and
-// discard is the only commit either flow makes. Push-CAS on origin arbitrates
-// concurrent writers; Mutate owns the bounded fast-forward retry (design D5 —
-// the retired REST path's org-keyed leaky bucket is not ported).
-// See docs/design/shared-volume-clone-architecture.md §6, §10.
+// The Workspace-port write primitive behind Save (annotated tag at the pinned
+// commit). Save creates no commit — the accepted draft is already on `main`.
+// Push-CAS on origin arbitrates concurrent writers; Mutate owns the bounded
+// fast-forward retry (design D5 — the retired REST path's org-keyed leaky
+// bucket is not ported). See docs/design/shared-volume-clone-architecture.md
+// §6, §10. (Discard's revert-commit primitive, revertSubtreeToTag, was
+// removed with DiscardRequirements/DiscardDesign — dead once the
+// requirements/design read+discard HTTP surface was removed.)
 
 import (
 	"context"
 	"errors"
-	"fmt"
 	"math/rand/v2"
-	"sort"
-	"strings"
 	"time"
 
 	"github.com/wso2/aep/aep-api/internal/feature/gitrepo"
@@ -98,67 +96,6 @@ func (s *artifactService) createAnnotatedTag(
 		err = attempt()
 	}
 	return err
-}
-
-// revertSubtreeToTag rewrites the `prefix` subtree on the default branch to
-// match its content at `tag` via one Mutate (the target files are staged as
-// writes; HEAD-subtree paths absent from the target are staged as deletes).
-// An empty `tag` reverts the subtree to empty (nothing approved). When the
-// subtree already matches the target the staged tree equals the base tree and
-// Mutate commits nothing. Returns the reverted bundle.
-func (s *artifactService) revertSubtreeToTag(
-	ctx context.Context,
-	ref gitrepo.RepoRef,
-	prefix string,
-	keep bundleFilter,
-	tag, message string,
-) (map[string]string, error) {
-	// Target subtree content at the tag — read once BEFORE the Mutate and
-	// treated as fixed input: the tag is immutable, so fn re-runs on a CAS
-	// retry stay correct. Every blob under prefix participates (no extension
-	// filter — the revert restores the subtree wholesale). Empty when there is
-	// no tag yet (revert-to-empty).
-	target := map[string]string{}
-	if tag != "" {
-		files, _, err := s.git.Workspace().ReadBundle(ctx, ref, "tags/"+tag, func(path string) bool {
-			return strings.HasPrefix(path, prefix)
-		})
-		if err != nil {
-			if errors.Is(err, gitrepo.ErrRefNotFound) {
-				return nil, fmt.Errorf("resolve tag %s: %w", tag, ErrArtifactNotFound)
-			}
-			return nil, fmt.Errorf("read bundle at tag %s: %w", tag, err)
-		}
-		target = files
-	}
-	// Deterministic staging order (map iteration is random; the resulting
-	// tree is identical either way, but ordered ops keep runs reproducible).
-	targetPaths := make([]string, 0, len(target))
-	for path := range target {
-		targetPaths = append(targetPaths, path)
-	}
-	sort.Strings(targetPaths)
-
-	author, committer := s.git.ResolveSaveIdentities(ref.Cred)
-	_, err := s.git.Workspace().Mutate(ctx, ref, func(tx gitrepo.Tx) error {
-		for _, path := range targetPaths {
-			tx.Write(path, []byte(target[path]))
-		}
-		// HEAD-subtree paths absent from the target are tombstoned.
-		return tx.Base().Walk(prefix, func(rel, _ string) error {
-			if _, keepIt := target[rel]; !keepIt {
-				tx.Delete(rel)
-			}
-			return nil
-		})
-	}, gitrepo.CommitOpts{Message: message, Author: author, Committer: committer})
-	if err != nil {
-		return nil, fmt.Errorf("revert commit: %w", err)
-	}
-
-	// Read the reverted bundle back through the workspace mount — the read
-	// fetches origin first, so the just-pushed revert is always visible.
-	return s.readBundleAtHead(ctx, ref, prefix, keep)
 }
 
 // jitterSleep waits for `base` with ±50% uniform jitter, respecting ctx
