--- services/aep-api/models/spec.go
@@ -1,29 +1,0 @@
-// Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
-//
-// WSO2 LLC. licenses this file to you under the Apache License,
-// Version 2.0 (the "License"); you may not use this file except
-// in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package models
-
-// RequirementsBundle describes the multi-document requirements set stored at
-// `specs/requirements/*.md`. The bundle is versioned together as `v<N>`
-// tags; each save produces one new tag covering the whole directory.
-type RequirementsBundle struct {
-	ProjectID         string            `json:"projectId"`
-	Files             map[string]string `json:"files"`
-	Status            string            `json:"status"` // "draft" | "approved"
-	Version           int               `json:"version,omitempty"`
-	Versions          []ArtifactVersion `json:"versions,omitempty"`
-	HasUnsavedChanges bool              `json:"hasUnsavedChanges,omitempty"`
-}
