--- services/aep-api/testdata/harvest/golden/get_design_bundle_at_tag.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "http://localhost:9090/api/v1/DesignBundle.json",
-  "files": {
-    "components/hello-api/openapi.yaml": "openapi: 3.0.3\ninfo:\n  title: Hello World API\n  description: A simple API that responds with a friendly greeting.\n  version: 1.0.0\nservers:\n  - url: /\npaths:\n  /health:\n    get:\n      summary: Health check\n      operationId: getHealth\n      responses:\n        '200':\n          description: Service is healthy\n          content:\n            text/plain:\n              schema:\n                type: string\n                example: OK\n  /hello:\n    get:\n      summary: Get a friendly greeting\n      operationId: getHello\n      parameters:\n        - name: name\n          in: query\n          required: false\n          description: Optional name to personalize the greeting.\n          schema:\n            type: string\n      responses:\n        '200':\n          description: A friendly greeting\n          content:\n            application/json:\n              schema:\n                type: object\n                properties:\n                  message:\n                    type: string\n                    example: Hello, World!\n                required:\n                  - message\n        '400':\n          description: Invalid request\n          content:\n            application/problem+json:\n              schema:\n                $ref: '#/components/schemas/Problem'\ncomponents:\n  schemas:\n    Problem:\n      type: object\n      properties:\n        type:\n          type: string\n          example: about:blank\n        title:\n          type: string\n          example: Bad Request\n        status:\n          type: integer\n          example: 400\n        detail:\n          type: string\n",
-    "design.md": "---\nskillsApplied:\n    - api-management\n    - go\n    - react-webapp\n    - thunder-authentication\n---\n\nA minimal single-service backend: a Go HTTP API that returns a friendly greeting. It exposes a public GET endpoint that replies \"Hello, World!\" by default, or \"Hello, [name]!\" when a name is supplied as a query parameter. No authentication, no persistence, no UI — just a health-checked greeting endpoint for developers to test connectivity.\n",
-    "components/hello-api/design.json": "{\n  \"name\": \"hello-api\",\n  \"type\": \"service\",\n  \"language\": \"Go\",\n  \"buildpack\": \"docker\",\n  \"appPath\": \"hello-api\",\n  \"entrypoint\": \"deployment/service\",\n  \"description\": \"Build a simple Go + net/http service on port 9090.\\n\\nEndpoints:\\n- GET /health — returns 200 \\\"OK\\\", no auth, cheap check.\\n- GET /hello — returns JSON {\\\"message\\\": \\\"Hello, World!\\\"}. If a query parameter `name` is provided (e.g. /hello?name=Alice), return {\\\"message\\\": \\\"Hello, Alice!\\\"} instead.\\n\\nThis is a public, unauthenticated API — do NOT add any /auth/* endpoints, JWT validation, or CORS middleware (there is no gateway auth requirement here since exposesAPI is omitted, and the platform's gateway config handles CORS if/when applied).\\n\\nNo persistence needed — this service is stateless, do NOT add a database, storage component, or scheduled/cron tasks.\\n\\nLayout: single main.go is fine for a service this small (see the `go` platform skill for the full recommended layout if you want to expand). Use stdlib net/http only — no external dependencies required, so no go.sum is expected.\\n\\nDockerfile: multi-stage build, `FROM golang:1.25-alpine AS builder`, CGO_ENABLED=0, slim `alpine:3.20` runtime, EXPOSE 9090.\\n\\nErrors (if any, e.g. bad query param) should use application/problem+json with type/title/status fields.\",\n  \"dependencies\": []\n}\n"
-  },
-  "design": {
-    "projectId": "hello-world-api",
-    "overview": "A minimal single-service backend: a Go HTTP API that returns a friendly greeting. It exposes a public GET endpoint that replies \"Hello, World!\" by default, or \"Hello, [name]!\" when a name is supplied as a query parameter. No authentication, no persistence, no UI — just a health-checked greeting endpoint for developers to test connectivity.",
-    "components": [
-      {
-        "name": "hello-api",
-        "componentType": "service",
-        "language": "Go",
-        "dependsOn": [],
-        "entrypoint": "deployment/service",
-        "buildpack": "docker",
-        "appPath": "hello-api",
-        "openAPISpec": "openapi: 3.0.3\ninfo:\n  title: Hello World API\n  description: A simple API that responds with a friendly greeting.\n  version: 1.0.0\nservers:\n  - url: /\npaths:\n  /health:\n    get:\n      summary: Health check\n      operationId: getHealth\n      responses:\n        '200':\n          description: Service is healthy\n          content:\n            text/plain:\n              schema:\n                type: string\n                example: OK\n  /hello:\n    get:\n      summary: Get a friendly greeting\n      operationId: getHello\n      parameters:\n        - name: name\n          in: query\n          required: false\n          description: Optional name to personalize the greeting.\n          schema:\n            type: string\n      responses:\n        '200':\n          description: A friendly greeting\n          content:\n            application/json:\n              schema:\n                type: object\n                properties:\n                  message:\n                    type: string\n                    example: Hello, World!\n                required:\n                  - message\n        '400':\n          description: Invalid request\n          content:\n            application/problem+json:\n              schema:\n                $ref: '#/components/schemas/Problem'\ncomponents:\n  schemas:\n    Problem:\n      type: object\n      properties:\n        type:\n          type: string\n          example: about:blank\n        title:\n          type: string\n          example: Bad Request\n        status:\n          type: integer\n          example: 400\n        detail:\n          type: string\n",
-        "componentAgentInstructions": "# hello-api\n\nBuild a simple Go + net/http service on port 9090.\n\nEndpoints:\n- GET /health — returns 200 \"OK\", no auth, cheap check.\n- GET /hello — returns JSON {\"message\": \"Hello, World!\"}. If a query parameter `name` is provided (e.g. /hello?name=Alice), return {\"message\": \"Hello, Alice!\"} instead.\n\nThis is a public, unauthenticated API — do NOT add any /auth/* endpoints, JWT validation, or CORS middleware (there is no gateway auth requirement here since exposesAPI is omitted, and the platform's gateway config handles CORS if/when applied).\n\nNo persistence needed — this service is stateless, do NOT add a database, storage component, or scheduled/cron tasks.\n\nLayout: single main.go is fine for a service this small (see the `go` platform skill for the full recommended layout if you want to expand). Use stdlib net/http only — no external dependencies required, so no go.sum is expected.\n\nDockerfile: multi-stage build, `FROM golang:1.25-alpine AS builder`, CGO_ENABLED=0, slim `alpine:3.20` runtime, EXPOSE 9090.\n\nErrors (if any, e.g. bad query param) should use application/problem+json with type/title/status fields."
-      }
-    ],
-    "status": "approved",
-    "version": 0,
-    "hasUnsavedChanges": false,
-    "sourceSpec": "v1"
-  }
-}
