--- services/aep-api/testdata/harvest/golden/get_design_versions.json
@@ -1,8 +1,0 @@
-[
-  {
-    "version": 1,
-    "tagName": "v1-1",
-    "commitHash": "d81ceeb71f46d9d585ead5b06a6c15b32d624d7d",
-    "sourceSpec": "v1"
-  }
-]
