--- services/aep-api/testdata/harvest/golden/get_requirements_versions.json
@@ -1,7 +1,0 @@
-[
-  {
-    "version": 1,
-    "tagName": "v1",
-    "commitHash": "f1582aff60d291d3081743c4649a20096b956ee0"
-  }
-]
